+++ features/support/requires.rb
@@ -5,6 +5,7 @@ class Requires
 	  require "providers/dss_associate"
 	  require "providers/jsender"
 	  require "providers/bootstrapped_provider"
+	  require "providers/juddi_soap_connector"
 	  require "providers/juddi_provider"
 	  require "test/orchestration_provider_registry"
 	  require "test/base_orchestration_provider"
